qcacld-3.0: Fix unnecessary logs in connect path
Fix unnecessary logs in connect path.

CRs-Fixed: 3016952

+++ umac/mlme/connection_mgr/dispatcher/inc/wlan_cm_roam_public_struct.h
@@ -344,7 +344,7 @@ struct rso_config {
 	struct reassoc_timer_ctx ctx;
 #endif
 	qdf_mutex_t cm_rso_lock;
-	uint8_t rsn_cap;
+	uint16_t rsn_cap;
 	uint8_t country_code[REG_ALPHA2_LEN + 1];
 	bool disable_hi_rssi;
 	bool roam_control_enable;
